Stack game using c
A and B are playing a game. In this game, both A and B are initially provided with two boxes
(one for each) which have a set of number (integers) blocks in each box. Number blocks are
arranged in a pattern that one block is kept over another (concept of Stack) and these blocks
can be only accessed from top end of the box. At the beginning, system will generate a
number set and A will get the number blocks in the same order while B gets it in the reverse
order (see the example below).
Now, both A and B have to take out the number at the open end of their own boxes
simultaneously.
System needs to generate the output as a series of 0,1 and 2 and then print the winner.
Once game is started, they have to follow below steps at each iteration until the game ends.
a. If the number taken out by A is bigger than B’s, A put away the number taken out. B
has to put back the number block into the box. System prints 2 on the screen.
b. If the number picked by A is smaller than B’s, B put away the number taken out. A
has to put back the number block into the box. System prints 1 on the screen.
c. If both have the same number, then both A and B can put away the number blocks
taken out. System prints 0 on the screen.
When one of them left with an empty box, it marks the end of the game. The winner is the
one who took all the number blocks out of the box.
Write a complete C program to implement this game. Count of the number set should be
taken from the keyboard. Accordingly, the system should generate the number set randomly
within the range 0 to 10. This number set should vary for different execution time, i.e. it
should not generate the same set every time the program executed.
Hint: Use the concept of Stack
Example: Input: 5
System generated number set: 2 9 1 6 7
A’s number set is 2 9 1 6 7 (7 is at the top of the box)
B’s number set is 7 6 1 9 2 (2 is at the top of the box)
Output: 2 2 1 1 0 2 1 1
B is the Winner